requirements.txt
requests>=2.28.0
openai>=2.0
python-dotenv>=1.0.0
flask>=3.0.0
# Pin httpx to avoid "proxies" kwarg incompatibility with openai (httpx 0.28+ removed it)
httpx>=0.24.0,<0.28.0
